third_party/blink/web_tests/external/wpt/css/css-values/sibling-function-invalidation-expected.txt

This is a testharness.js-based test.
[FAIL] 5th sibling after removal
  assert_equals: expected 50 but got 60
[FAIL] 5 siblings after removal
  assert_equals: expected 50 but got 60
Harness: the test ran to completion.
